backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.2
google-generativeai==0.4.1
openai==1.14.0
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
aiofiles==23.2.1
pillow==10.3.0
httpx==0.27.0
scikit-learn==1.4.2
slowapi==0.1.9
jinja2==3.1.3
email-validator==2.1.0
